+++ http-gateway/web/src/containers/things/utils.js
@@ -1,3 +1,4 @@
+import { time } from 'units-converter'
 import { getApiErrorMessage } from '@/common/utils'
 import { showErrorToast, showWarningToast } from '@/components/toast'
 import { compareIgnoreCase } from '@/components/table/utils'
@@ -9,8 +10,12 @@ import {
   THINGS_RESOURCE_REGISTRATION_WS_KEY,
   THINGS_RESOURCE_UPDATE_WS_KEY,
   shadowSynchronizationStates,
+  commandTimeoutUnits,
+  MINIMAL_TTL_VALUE_MS,
 } from './constants'
 import { messages as t } from './things-i18n'
+
+const { INFINITE, MS, NS } = commandTimeoutUnits
 
 // Returns the extension for resources API for the selected interface
 export const interfaceGetParam = currentInterface =>
@@ -233,6 +238,40 @@ export const getLastPartOfAResourceHref = href => {
   const values = href.split('/').filter(t => t)
   return values[values.length - 1]
 }
+
+// Converts a value to ns (if the unit is Infinite, it defaults to ns)
+export const convertValueToNs = (value, unit) =>
+  +time(value)
+    .from(unit === INFINITE ? NS : unit)
+    .to(NS)
+    .value.toFixed(0)
+
+// Converts a value from a given unit to a provided unit (if the unit is Infinite, it defaults to ns)
+export const convertValueFromTo = (value, unitFrom, unitTo) =>
+  time(value)
+    .from(unitFrom === INFINITE ? NS : unitFrom)
+    .to(unitTo === INFINITE ? NS : unitTo).value
+
+// Normalizes a given value to a fixed float number
+export const normalizeToFixedFloatValue = value => +value.toFixed(5)
+
+// Return true if there is a command timeout error based on the provided value and unit
+export const hasCommandTimeoutError = (value, unit) => {
+  const baseUnit = unit === INFINITE ? NS : unit
+
+  const valueMs = time(value)
+    .from(baseUnit)
+    .to(MS).value
+
+  if (valueMs < MINIMAL_TTL_VALUE_MS && value !== 0) {
+    return true
+  }
+
+  return false
+}
+
+export const convertAndNormalizeValueFromTo = (value, unitFrom, unitTo) =>
+  normalizeToFixedFloatValue(convertValueFromTo(value, unitFrom, unitTo))
 
 // Returns a boolean wether the shadow synchronization is enabled
 export const shadowSynchronizationEnabled = shadowSynchronization =>
